GH-2560 fix Model equals/hashcode by no longer using isomorphism
- added regression tests
- minor refactor of DynamicModel to inherit from AbstractSet
- minor refactor of testsuite (renamed class, updated javadoc, fixed
copyright header)
