+++ compliance/model/src/test/java/org/eclipse/rdf4j/model/DynamicModelTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015 Eclipse RDF4J contributors, Aduna, and others.
+ * Copyright (c) 2020 Eclipse RDF4J contributors.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Distribution License v1.0
  * which accompanies this distribution, and is available at
@@ -10,7 +10,7 @@ package org.eclipse.rdf4j.model;
 import org.eclipse.rdf4j.model.impl.DynamicModel;
 import org.eclipse.rdf4j.model.impl.LinkedHashModelFactory;
 
-public class DynamicModelTest extends AbstractModelTest {
+public class DynamicModelTest extends ModelTest {
 
 	@Override
 	protected Model getNewModel() {
